Sample EMTs within LightSamplerBase
